+++ src/Utility/font.cpp
@@ -4,7 +4,7 @@ int Fonts::init(Bundle* resources) {
     
     //Load font data file
     font_dat = resources->load_raw_data("resources/fonts/font.dat");
-    if (font_dat.size() != 273) { printf("Font data size wrong! at %d",font_dat.size()); return 1;}
+    if (font_dat.size() != 273) { printf("Font data size wrong! at %d",(int)font_dat.size()); return 1;}
     //Load texture atlas
     //Eventually add this to angel
     std::vector<unsigned char> image;
